Remove unnecessary tts transformer parts.
