--- nemo/collections/tts/parts/transformer/conv_transpose.py
@@ -1,55 +1,0 @@
-# Copyright 2020 NVIDIA. All Rights Reserved.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-from torch import nn
-
-
-class Conv(nn.Module):
-    """
-    Convolution Module
-    """
-
-    def __init__(
-        self, in_channels, out_channels, kernel_size=1, stride=1, padding=0, dilation=1, bias=True, w_init='relu'
-    ):
-        """
-        :param in_channels: dimension of input
-        :param out_channels: dimension of output
-        :param kernel_size: size of kernel
-        :param stride: size of stride
-        :param padding: size of padding
-        :param dilation: dilation rate
-        :param bias: boolean. if True, bias is included.
-        :param w_init: str. weight inits with xavier initialization.
-        """
-        super(Conv, self).__init__()
-
-        self.conv = nn.Conv1d(
-            in_channels,
-            out_channels,
-            kernel_size=kernel_size,
-            stride=stride,
-            padding=padding,
-            dilation=dilation,
-            bias=bias,
-        )
-
-        nn.init.xavier_uniform_(self.conv.weight, gain=nn.init.calculate_gain(w_init))
-
-    def forward(self, x):
-        x = x.contiguous().transpose(1, 2)
-        x = self.conv(x)
-        x = x.contiguous().transpose(1, 2)
-
-        return x
