+++ nemo/collections/tts/parts/transformer/layers.py
@@ -12,230 +12,113 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from collections import OrderedDict
-
+import numpy as np
 import torch
+import torch.nn.functional as F
 from torch import nn
-from torch.nn import functional as F
-
-from nemo.collections.tts.parts.transformer import sublayers
 
 
-class Linear(nn.Module):
-    """
-    Linear Module
-    """
+class ScaledDotProductAttention(nn.Module):
+    """Scaled Dot-Product Attention."""
 
-    def __init__(self, in_dim, out_dim, bias=True, w_init='linear'):
-        """
-        :param in_dim: dimension of input
-        :param out_dim: dimension of output
-        :param bias: boolean. if True, bias is included.
-        :param w_init: str. weight inits with xavier initialization.
-        """
-        super(Linear, self).__init__()
-        self.linear_layer = nn.Linear(in_dim, out_dim, bias=bias)
+    def __init__(self, temperature, attn_dropout=0.1):
+        super().__init__()
+        self.temperature = temperature
+        self.dropout = nn.Dropout(attn_dropout)
+        self.softmax = nn.Softmax(dim=2)
 
-        nn.init.xavier_uniform_(self.linear_layer.weight, gain=nn.init.calculate_gain(w_init))
+    def forward(self, q, k, v, mask=None):
 
-    def forward(self, x):
-        return self.linear_layer(x)
+        attn = torch.bmm(q, k.transpose(1, 2))
+        attn = attn / self.temperature
 
+        if mask is not None:
+            attn = attn.masked_fill(mask, -1e-9)
 
-class PreNet(nn.Module):
-    """
-    Pre Net before passing through the network
-    """
+        attn = self.softmax(attn)
+        attn = self.dropout(attn)
+        output = torch.bmm(attn, v)
 
-    def __init__(self, input_size, hidden_size, output_size, p=0.5):
-        """
-        :param input_size: dimension of input
-        :param hidden_size: dimension of hidden unit
-        :param output_size: dimension of output
-        """
-        super(PreNet, self).__init__()
-        self.input_size = input_size
-        self.output_size = output_size
-        self.hidden_size = hidden_size
-        self.layer = nn.Sequential(
-            OrderedDict(
-                [
-                    ('fc1', Linear(self.input_size, self.hidden_size)),
-                    ('relu1', nn.ReLU()),
-                    ('dropout1', nn.Dropout(p)),
-                    ('fc2', Linear(self.hidden_size, self.output_size)),
-                    ('relu2', nn.ReLU()),
-                    ('dropout2', nn.Dropout(p)),
-                ]
-            )
-        )
-
-    def forward(self, input_):
-
-        out = self.layer(input_)
-
-        return out
+        return output, attn
 
 
-class Conv(nn.Module):
-    """
-    Convolution Module
-    """
+class MultiHeadAttention(nn.Module):
+    """Multi-Head Attention module."""
 
-    def __init__(
-        self, in_channels, out_channels, kernel_size=1, stride=1, padding=0, dilation=1, bias=True, w_init='linear'
-    ):
-        """
-        :param in_channels: dimension of input
-        :param out_channels: dimension of output
-        :param kernel_size: size of kernel
-        :param stride: size of stride
-        :param padding: size of padding
-        :param dilation: dilation rate
-        :param bias: boolean. if True, bias is included.
-        :param w_init: str. weight inits with xavier initialization.
-        """
-        super(Conv, self).__init__()
+    def __init__(self, n_head, d_model, d_k, d_v, dropout=0.1):
+        super().__init__()
 
-        self.conv = nn.Conv1d(
-            in_channels,
-            out_channels,
-            kernel_size=kernel_size,
-            stride=stride,
-            padding=padding,
-            dilation=dilation,
-            bias=bias,
-        )
+        self.n_head = n_head
+        self.d_k = d_k
+        self.d_v = d_v
 
-        nn.init.xavier_uniform_(self.conv.weight, gain=nn.init.calculate_gain(w_init))
+        self.w_qs = nn.Linear(d_model, n_head * d_k)
+        self.w_ks = nn.Linear(d_model, n_head * d_k)
+        self.w_vs = nn.Linear(d_model, n_head * d_v)
+        nn.init.normal_(self.w_qs.weight, mean=0, std=np.sqrt(2.0 / (d_model + d_k)))
+        nn.init.normal_(self.w_ks.weight, mean=0, std=np.sqrt(2.0 / (d_model + d_k)))
+        nn.init.normal_(self.w_vs.weight, mean=0, std=np.sqrt(2.0 / (d_model + d_v)))
 
-    def forward(self, x):
-        x = self.conv(x)
-        return x
+        self.attention = ScaledDotProductAttention(temperature=np.power(d_k, 0.5))
+        self.layer_norm = nn.LayerNorm(d_model)
 
+        self.fc = nn.Linear(n_head * d_v, d_model)
+        nn.init.xavier_normal_(self.fc.weight)
 
-class FFTBlock(torch.nn.Module):
-    """FFT Block"""
+        self.dropout = nn.Dropout(dropout)
 
-    def __init__(self, d_model, d_inner, n_head, d_k, d_v, fft_conv1d_kernel, fft_conv1d_padding, dropout=0.1):
-        super(FFTBlock, self).__init__()
-        self.slf_attn = sublayers.MultiHeadAttention(n_head, d_model, d_k, d_v, dropout=dropout)
-        self.pos_ffn = sublayers.PositionwiseFeedForward(
-            d_model,
-            d_inner,
-            fft_conv1d_kernel=fft_conv1d_kernel,
-            fft_conv1d_padding=fft_conv1d_padding,
-            dropout=dropout,
-        )
+    def forward(self, q, k, v, mask=None):
 
-    def forward(self, enc_input, non_pad_mask=None, slf_attn_mask=None):
-        enc_output, enc_slf_attn = self.slf_attn(enc_input, enc_input, enc_input, mask=slf_attn_mask)
-        enc_output *= non_pad_mask
+        d_k, d_v, n_head = self.d_k, self.d_v, self.n_head
 
-        enc_output = self.pos_ffn(enc_output)
-        enc_output *= non_pad_mask
+        sz_b, len_q, _ = q.size()
+        sz_b, len_k, _ = k.size()
+        sz_b, len_v, _ = v.size()
 
-        return enc_output, enc_slf_attn
+        residual = q
+
+        q = self.w_qs(q).view(sz_b, len_q, n_head, d_k)
+        k = self.w_ks(k).view(sz_b, len_k, n_head, d_k)
+        v = self.w_vs(v).view(sz_b, len_v, n_head, d_v)
+
+        q = q.permute(2, 0, 1, 3).contiguous().view(sz_b * n_head, len_q, d_k)  # (n*b) x lq x dk
+        k = k.permute(2, 0, 1, 3).contiguous().view(sz_b * n_head, len_k, d_k)  # (n*b) x lk x dk
+        v = v.permute(2, 0, 1, 3).contiguous().view(sz_b * n_head, len_v, d_v)  # (n*b) x lv x dv
+
+        mask = mask.repeat(n_head, 1, 1)  # (n*b) x .. x ..
+        output, attn = self.attention(q, k, v, mask=mask)
+
+        output = output.view(n_head, sz_b, len_q, d_v)
+        output = output.permute(1, 2, 0, 3).contiguous().view(sz_b, len_q, n_head * d_v)  # b x lq x (n*dv)
+
+        output = self.dropout(self.fc(output))
+        output = self.layer_norm(output + residual)
+
+        return output, attn
 
 
-class ConvNorm(torch.nn.Module):
-    def __init__(
-        self,
-        in_channels,
-        out_channels,
-        kernel_size=1,
-        stride=1,
-        padding=None,
-        dilation=1,
-        bias=True,
-        w_init_gain='linear',
-    ):
-        super(ConvNorm, self).__init__()
+class PositionwiseFeedForward(nn.Module):
+    """A two-feed-forward-layer module."""
 
-        if padding is None:
-            assert kernel_size % 2 == 1
-            padding = int(dilation * (kernel_size - 1) / 2)
+    def __init__(self, d_in, d_hid, fft_conv1d_kernel, fft_conv1d_padding, dropout=0.1):
+        super().__init__()
 
-        self.conv = torch.nn.Conv1d(
-            in_channels,
-            out_channels,
-            kernel_size=kernel_size,
-            stride=stride,
-            padding=padding,
-            dilation=dilation,
-            bias=bias,
-        )
+        # Use Conv1D
 
-        torch.nn.init.xavier_uniform_(self.conv.weight, gain=torch.nn.init.calculate_gain(w_init_gain))
+        # position-wise
+        self.w_1 = nn.Conv1d(d_in, d_hid, kernel_size=fft_conv1d_kernel, padding=fft_conv1d_padding)
+        # position-wise
+        self.w_2 = nn.Conv1d(d_hid, d_in, kernel_size=fft_conv1d_kernel, padding=fft_conv1d_padding)
 
-    def forward(self, signal):
-        conv_signal = self.conv(signal)
-
-        return conv_signal
-
-
-class PostNet(nn.Module):
-    """
-    PostNet: Five 1-d convolution with 512 channels and kernel size 5
-    """
-
-    def __init__(self, n_mel_channels=80, postnet_embedding_dim=512, postnet_kernel_size=5, postnet_n_convolutions=5):
-
-        super(PostNet, self).__init__()
-        self.convolutions = nn.ModuleList()
-
-        self.convolutions.append(
-            nn.Sequential(
-                ConvNorm(
-                    n_mel_channels,
-                    postnet_embedding_dim,
-                    kernel_size=postnet_kernel_size,
-                    stride=1,
-                    padding=int((postnet_kernel_size - 1) / 2),
-                    dilation=1,
-                    w_init_gain='tanh',
-                ),
-                nn.BatchNorm1d(postnet_embedding_dim),
-            )
-        )
-
-        for i in range(1, postnet_n_convolutions - 1):
-            self.convolutions.append(
-                nn.Sequential(
-                    ConvNorm(
-                        postnet_embedding_dim,
-                        postnet_embedding_dim,
-                        kernel_size=postnet_kernel_size,
-                        stride=1,
-                        padding=int((postnet_kernel_size - 1) / 2),
-                        dilation=1,
-                        w_init_gain='tanh',
-                    ),
-                    nn.BatchNorm1d(postnet_embedding_dim),
-                )
-            )
-
-        self.convolutions.append(
-            nn.Sequential(
-                ConvNorm(
-                    postnet_embedding_dim,
-                    n_mel_channels,
-                    kernel_size=postnet_kernel_size,
-                    stride=1,
-                    padding=int((postnet_kernel_size - 1) / 2),
-                    dilation=1,
-                    w_init_gain='linear',
-                ),
-                nn.BatchNorm1d(n_mel_channels),
-            )
-        )
+        self.layer_norm = nn.LayerNorm(d_in)
+        self.dropout = nn.Dropout(dropout)
 
     def forward(self, x):
-        x = x.contiguous().transpose(1, 2)
+        residual = x
+        output = x.transpose(1, 2)
+        output = self.w_2(F.relu(self.w_1(output)))
+        output = output.transpose(1, 2)
+        output = self.dropout(output)
+        output = self.layer_norm(output + residual)
 
-        for i in range(len(self.convolutions) - 1):
-            x = F.dropout(torch.tanh(self.convolutions[i](x)), 0.5, self.training)
-        x = F.dropout(self.convolutions[-1](x), 0.5, self.training)
-
-        x = x.contiguous().transpose(1, 2)
-        return x
+        return output
